refactor: rename 'pip' utils to 'packages' utils
Since we now have both pip and uv backends, the pip naming doesn't
make much sense anymore.
