+++ src/sophys/common/utils/packages.py
